# Conflicts:
#	buildconfig/appcompile/compile-customerapp.gradle
#	buildconfig/appcompile/dynamic-feature-customerapp.cfg
#	buildconfig/appcompile/dynamic-feature-customerapp_pro.cfg
#	buildconfig/dependencies/dependencies.gradle
#	buildconfig/dependencies/dependency-features.gradle
#	customerapp/proguard-rules.txt
#	customerapp/src/main/res/values/feature_names.xml
#	customerapp_pro/src/main/res/values/feature_names.xml
#	features/home/searchbar/src/main/java/com/tokopedia/searchbar/navigation_component/NavToolbar.kt
#	features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/data/model/ProductInfoP2Data.kt
#	features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/data/model/ProductInfoP2UiData.kt
#	features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/data/util/DynamicProductDetailMapper.kt
#	features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/data/util/DynamicProductDetailTracking.kt
#	features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/usecase/GetProductInfoP2DataUseCase.kt
#	features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/view/adapter/factory/DynamicProductDetailAdapterFactory.kt
#	features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/view/adapter/factory/DynamicProductDetailAdapterFactoryImpl.kt
#	features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/view/fragment/DynamicProductDetailFragment.kt
#	features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/view/listener/DynamicProductDetailListener.kt
#	features/merchant/product_detail/src/main/java/com/tokopedia/product/detail/view/util/PdpUiUpdater.kt
#	features/merchant/product_detail/src/test/java/com/tokopedia/product/util/ProductDetailTestUtil.kt
#	features/merchant/product_detail/src/test/java/com/tokopedia/product/viewmodel/DynamicProductDetailViewModelTest.kt
#	features/merchant/product_detail/src/test/resources/json/gql_get_pdp_p2_data.json
#	features/merchant/product_detail_common/src/main/java/com/tokopedia/product/detail/common/ProductViewHelper.kt
#	features/merchant/product_detail_common/src/main/java/com/tokopedia/product/detail/common/SingleClickUtil.kt
#	features/merchant/product_detail_common/src/main/java/com/tokopedia/product/detail/common/data/model/pdplayout/CampaignModular.kt
#	libraries/applink/src/main/java/com/tokopedia/applink/DeeplinkDFMapper.kt
#	libraries/applink/src/main/java/com/tokopedia/applink/DeeplinkMapper.kt

+++ features/content/kol_common/src/main/java/com/tokopedia/kolcommon/view/viewmodel/LikeKolViewModel.kt
@@ -1,12 +1,10 @@
 package com.tokopedia.kolcommon.view.viewmodel
 
-import com.tokopedia.kolcommon.data.pojo.FollowKolDomain
-
 /**
  * @author by yoasfs on 2019-12-05
  */
 data class LikeKolViewModel (
-        var id: Int = 0,
+        var id: Long = 0,
         var rowNumber: Int = 0,
         var errorMessage: String = "",
         var isSuccess: Boolean = false
